+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 * Added `topicoptions.WithReaderOnStopPartitionSession` to invoke the user callback when the server stops a reader partition session
 
+## v3.135.13
+* Optimized the `internal/stack.FunctionID` for decreased allocations and CPU usage
+
+## v3.135.12
+* Optimized the `internal/meta.TraceID` generation of `x-ydb-trace-id` gRPC header
+
 ## v3.135.11
 * The `trace.DatabaseSQLConnExecStartInfo.IdleTime` and `trace.DatabaseSQLConnQueryStartInfo.IdleTime` fields have been marked as deprecated and will always be zero from now.
 * Default connection last usage tracking has been changed to "no tracking", with real-time tracking enabled only if a connection's time-to-live (TTL) has been defined.